\name{soft.threshhold}
\alias{soft.threshhold}
\title{Soft-thresholding...}
\usage{soft.threshhold(x, lambda=1)}
\description{Soft-thresholding}
\details{Apply the soft-threshold function to a vector.}
\value{The vector of observations after applying the
soft-thresholding.}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}
\arguments{\item{x}{A vector of numeric data.}
\item{lambda}{The largest absolute value that will be mapped to
zero.}}
