\name{zelig2gamma.mixed}
\alias{zelig2gamma.mixed}
\title{Interface between the Zelig Model gamma...}
\usage{zelig2gamma.mixed(formula, ..., data)}
\description{Interface between the Zelig Model gamma.mixed and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{}
\item{data}{a data.frame}}
