\name{zelig2logit.mixed}
\alias{zelig2logit.mixed}
\title{Interface between the Zelig Model logit...}
\usage{zelig2logit.mixed(formula, ..., data)}
\description{Interface between the Zelig Model logit.mixed and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{}
\item{data}{a data.frame}}
