\name{zelig2ls.mixed}
\alias{zelig2ls.mixed}
\title{Interface between the Zelig Model ls...}
\usage{zelig2ls.mixed(formula, ..., data)}
\description{Interface between the Zelig Model ls.mixed and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{}
\item{data}{a data.frame}}
